+++ miden-crypto/src/lib.rs
@@ -6,9 +6,9 @@ extern crate alloc;
 #[cfg(feature = "std")]
 extern crate std;
 
+pub mod aead;
 pub mod dsa;
 pub mod ecdh;
-pub mod encryption;
 pub mod hash;
 pub mod merkle;
 pub mod rand;